\name{Microgen.read}
\alias{Microgen.read}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Process Sheﬃeld Solar - Microgen data
}
\description{
Process, format, and clean Sheﬃeld Solar - Microgen data
}
\usage{
Microgen.read(root_data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{root_data}{
  data directory
  }
}

\value{
  data frame containing time series of different locations' PV power data
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
C. Feng, D. Yang, B.-M. Hodge, and J. Zhang
}
